enterprise_rag_system/requirements.txt
# ============================================================
# Enterprise Knowledge Intelligence System — Dependencies
# ============================================================

# ------------------------------------
# Web Framework & Server
# ------------------------------------
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9      # Required for FastAPI file uploads

# ------------------------------------
# Data Validation
# ------------------------------------
pydantic>=2.6.0

# ------------------------------------
# Document Parsing
# ------------------------------------
pdfplumber>=0.11.0
pypdf>=4.1.0

# ------------------------------------
# Embeddings
# ------------------------------------
sentence-transformers>=3.0.0

# ------------------------------------
# Vector Store
# ------------------------------------
faiss-cpu>=1.8.0
# Use faiss-gpu instead if you have a CUDA-capable GPU:
# faiss-gpu>=1.7.4

# ------------------------------------
# Local LLM — HuggingFace Transformers
# ------------------------------------
torch>=2.2.0
transformers>=4.40.0
accelerate>=0.29.0
# Optional: 4-bit quantisation (requires CUDA GPU)
# bitsandbytes>=0.43.0

# ------------------------------------
# Numerical Computing
# ------------------------------------
numpy>=1.26.0

# ------------------------------------
# Testing
# ------------------------------------
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0                # Async HTTP client for FastAPI TestClient
